backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
google-generativeai==0.4.2
httpx==0.26.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
pydantic==2.6.1
pydantic-settings==2.1.0
